+++ doc/design/colors.js
@@ -10,6 +10,7 @@ const sass = {
   'background-3': '#f6f6f6', // Light Grey
   'line-1': '#eee', // Semi Light Grey
   'line-2': '#d8d8d8', // Grey
+  'line-3': '#dbdbdb', // Grey
 
   'primary-1': '#19b4fa', // Primary
   'primary-2': '#05a8e6', // Semi Dark Primary
@@ -30,10 +31,14 @@ const sass = {
   'tertiary-3': '#4ae191',
 
   'valid': '#61d079', // Green
+  'valid1': 'rgba(97, 208, 121, .1)', // Green Light
 
   'error': '#ff0046', // Red
   'error-2': '#ffe5ec', // Light Red
   'error-3': '#ffb2c7', // Semi Light Red
+
+  'orange': '#ff7800',
+  'orange1': 'rgba(255, 130, 15, .1)', // Orange light
 
   'red-light-mdc': '#ffebe0',
   'red-mdc': '#ff0000',